+++ docs/changelog.md
@@ -23,6 +23,7 @@ Added:
 - Add message about importer thread being done.
 - Add `inspect-photo` entrypoint for debugging.
 - Add some debug output when `idxmax` cannot be taken. ([GH-307](https://github.com/martin-ueding/geo-activity-playground/issues/307))
+- Offer heatmap overlay in explorer tile map.
 
 Changed: